+++ src/fkooman/VPN/Server/AclInterface.php
@@ -24,7 +24,8 @@ interface AclInterface
      *
      * @param string userId the userID of the user to request the groups of
      *
-     * @return array the groups as an array, empty array if no groups
+     * @return array the groups as an array containing the keys "id" and 
+     *               "displayName", empty array if no groups are available for this user
      */
     public function getGroups($userId);
 }