+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 8.2.0 (...)
+- redo group support to have identifiers and display names instead of just 
+  identifiers and create mapping in the pool configuration
+- remove RemoteAcl for now
+
 ## 8.1.0 (2016-07-09)
 - remove the CRL support, only support disabling configurations